webmastered/requirements.txt
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
cloudflare==2.8.15
cryptography==35.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.4
django-cogwheels==0.3
django-comments-xtd==2.9.2
django-compressor==2.4.1
django-contrib-comments==2.1.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-filter==21.1
django-libsass==0.8
django-modelcluster==5.2
django-oauth-toolkit==1.5.0
django-redis==5.1.0
django-tables2==2.4.0
django-taggit==1.5.1
django-treebeard==4.5.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docutils==0.17.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jsonlines==2.0.0
jsonpickle==2.0.0
jwcrypto==1.0
l18n==2021.3
libsass==0.21.0
numpy==1.21.1
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.0
Pillow==9.0.0
plotly==5.1.0
psutil==5.8.0
psycopg2==2.9.1
pycountry==20.7.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-digitalocean==1.16.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
rcssmin==1.0.6
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
sentry-sdk==1.4.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stripe==2.63.0
tablib==3.1.0
telepath==0.2
tenacity==8.0.1
Unidecode==1.2.0
urllib3==1.26.7
wagtail==2.15.1
wagtail-color-panel==1.2.0
wagtailmenus==3.0.2
wagtailstreamforms==3.15
webencodings==0.5.1
Willow==1.4
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
